src/api/requirements.txt
Django==4.0
djangorestframework==3.12.4
psycopg2==2.8.6
daphne==3.0.2
python-dateutil==2.8.2
django-filter==21.1
django-cors-headers==3.10.0
celery==5.2.3
redis==4.1.0
djangorestframework-simplejwt==5.0.0
pandas==1.4.1
channels==3.0.4
channels-redis==3.3.1 